+++ main.py
@@ -3,11 +3,11 @@ import random
 import time
 import os
 from datetime import datetime
-from pathlib import Path
 from ui_helpers import draw_text, draw_wrapped_text, button, TextInputBox
 from data_loader import load_questions
 from utils import calculate_similarity, get_documents_folder
 from sound_manager import SoundManager
+from game_logic import TypingGame
 
 # Initialize Pygame and the mixer for sound.
 pygame.init()
@@ -19,6 +19,8 @@ clock = pygame.time.Clock()
 
 # sound_manager instance
 sound_manager = SoundManager()
+
+game = TypingGame(sound_manager, font, screen)
 
 # Default game settings
 default_game_length = 180  # seconds
@@ -35,389 +37,24 @@ GAME_OVER = "game_over"
 NORMAL_MODE = "normal"
 CLEAR_MODE = "clear"
 
-# Main Game Class
-class TypingGame:
-    def __init__(self, sound_manager):
-        self.state = MENU
-        self.pause_start = None  # Timestamp captured on pause
-        self.questions = load_questions()
-        self.available_questions = []  # For Clear Mode - questions not yet answered correctly
-        self.game_mode = NORMAL_MODE  # Default to normal mode
-        self.score = 0
-        self.start_time = None
-        self.question_timer = 0
-        self.current_question = None
-        self.user_input = ""
-        self.game_length = default_game_length
-        self.question_time = default_question_time
-        self.time_left = self.game_length
-        self.feedback = ""
-        self.feedback_color = (255, 255, 255)
-        self.feedback_timer = 0
-        self.last_question_answer = ""  # Previous question's answer
-        self.last_question_info = ""  # Previous question's extra info
-        self.learning_mode = False  # Learning mode off by default
-        self.sound_enabled = True  # Sound on by default
-        self.correct_streak = 0  # For consecutive correct answers
-        self.questions_answered = 0  # Total questions answered
-        self.questions_correct = 0  # Total questions answered correctly
-        self.clear_mode_correct = 0  # Track correct answers in Clear Mode
-        self.wrong_answers = []  # Track wrong answers for file output
-        self.wrong_answers_file = None  # Path to saved wrong answers file
-        self.sound_manager = sound_manager
-
-        # Create text input boxes for settings.
-        self.input_box_game_length = TextInputBox(300, 300, 200, 40, str(default_game_length), font)
-        self.input_box_question_time = TextInputBox(300, 370, 200, 40, str(default_question_time), font)
-
-    def pause(self):
-        if self.state == PLAYING:
-            self.state = PAUSED
-            self.pause_start = time.time()
-
-    def resume(self):
-        if self.state == PAUSED:
-            delta = time.time() - self.pause_start
-            # shift every running timer forward by the time we were paused
-            self.start_time += delta  # overall game clock
-            self.question_timer += delta  # current‑question deadline
-            self.feedback_timer += delta  # feedback visibility
-            self.state = PLAYING
-
-    def reset_game(self):
-        try:
-            self.game_length = int(self.input_box_game_length.text)
-        except ValueError:
-            self.game_length = default_game_length
-        try:
-            self.question_time = int(self.input_box_question_time.text)
-        except ValueError:
-            self.question_time = default_question_time
-
-        # Initialize available questions for Clear Mode
-        if self.game_mode == CLEAR_MODE:
-            self.available_questions = self.questions.copy()
-        
-        self.score = 0
-        self.start_time = time.time()
-        self.time_left = self.game_length
-        self.user_input = ""
-        self.feedback = ""
-        self.feedback_timer = 0
-        self.last_question_answer = ""
-        self.last_question_info = ""
-        self.correct_streak = 0  # Reset streak at game start.
-        self.questions_answered = 0  # Reset question counters
-        self.questions_correct = 0
-        self.clear_mode_correct = 0  # Reset Clear Mode counter
-        self.wrong_answers = []  # Reset wrong answers list
-        self.next_question()
-        self.state = PLAYING
-
-    def save_wrong_answers(self):
-        """Save wrong answers to a text file in Documents folder"""
-        if not self.wrong_answers:
-            return None
-        
-        # Generate filename with date and time
-        now = datetime.now()
-        date_str = now.strftime('%Y%m%d')
-        time_str = now.strftime('%H%M%S')
-        filename = f'WrongAnswers{date_str}{time_str}.txt'
-        
-        # Get Documents folder path
-        docs_folder = get_documents_folder()
-
-        # Create Documents folder if it doesn't exist
-        docs_folder.mkdir(exist_ok=True)
-        
-        filepath = docs_folder / filename
-        
-        try:
-            with open(filepath, 'w', encoding='utf-8') as f:
-                f.write('Wrong Answers Report\n')
-                f.write('=' * 50 + '\n\n')
-                
-                for i, entry in enumerate(self.wrong_answers, 1):
-                    f.write(f'Question {i}:\n')
-                    f.write(f'Q: {entry["question"]}\n')
-                    f.write(f'Correct Answer: {entry["correct_answer"]}\n')
-                    f.write(f'Your Answer: {entry["user_answer"]}\n')
-                    if entry['extra_info']:
-                        f.write(f'Extra Info: {entry["extra_info"]}\n')
-                    f.write('\n' + '-' * 30 + '\n\n')
-                
-                f.write(f'Total Wrong Answers: {len(self.wrong_answers)}\n')
-            
-            return str(filepath)
-        except Exception as e:
-            print(f'Error saving wrong answers: {e}')
-            return None
-
-    def next_question(self):
-        # Store the previous question's info.
-        if self.current_question is not None:
-            self.last_question_answer = self.current_question[1]
-            self.last_question_info = self.current_question[2]
-        
-        # Choose question based on game mode
-        if self.game_mode == CLEAR_MODE:
-            if not self.available_questions:
-                # All questions completed in Clear Mode
-                self.wrong_answers_file = self.save_wrong_answers()
-                self.state = GAME_OVER
-                return
-            # Choose from available questions
-            if len(self.available_questions) > 1:
-                new_question = random.choice(self.available_questions)
-                while self.current_question is not None and new_question == self.current_question:
-                    new_question = random.choice(self.available_questions)
-                self.current_question = new_question
-            else:
-                self.current_question = self.available_questions[0]
-        else:
-            # Normal mode - prevent the same question twice in a row.
-            if len(self.questions) > 1:
-                new_question = random.choice(self.questions)
-                while self.current_question is not None and new_question == self.current_question:
-                    new_question = random.choice(self.questions)
-                self.current_question = new_question
-            else:
-                self.current_question = random.choice(self.questions)
-        
-        self.question_timer = time.time() + self.question_time
-        self.user_input = ""
-        self.feedback_timer = time.time() + self.question_time
-
-    def draw_question_timer_bar(self):
-        if self.state == PLAYING and self.current_question:
-            total = self.question_time
-            remaining = max(0, self.question_timer - time.time())
-            fraction = remaining / total
-            bar_width = int(710 * fraction)
-            pygame.draw.rect(screen, (100, 100, 100), (40, 170, 710, 10))
-            pygame.draw.rect(screen, (0, 200, 0), (40, 170, bar_width, 10))
-
-    def draw_all_information(self):
-        if self.feedback and time.time() < self.feedback_timer:
-            feedback_surface = font.render(self.feedback, True, self.feedback_color)
-            feedback_rect = feedback_surface.get_rect(center=(400, 350))
-            screen.blit(feedback_surface, feedback_rect)
-            if self.learning_mode:
-                answer_text = f"Answer: {self.last_question_answer}"
-                answer_surface = font.render(answer_text, True, (255, 255, 0))
-                answer_rect = answer_surface.get_rect(center=(400, 390))
-                screen.blit(answer_surface, answer_rect)
-                if self.last_question_info.strip():
-                    info_text = f"Info: {self.last_question_info}"
-                    draw_wrapped_text(screen, info_text, (50, 430), font, color=(200, 200, 0), max_width=700)
-        self.draw_question_timer_bar()
-        draw_text(screen, f"Time Left: {self.time_left}s", (10, 10), font)
-        
-        # Show questions remaining in Clear Mode
-        if self.game_mode == CLEAR_MODE:
-            remaining_text = f"Questions Remaining: {len(self.available_questions)}"
-            draw_text(screen, remaining_text, (10, 40), font)
-            correct_count_text = f"Correct Questions: {self.clear_mode_correct}"
-            draw_text(screen, correct_count_text, (10, 70), font)
-        
-        draw_text(screen, f"Score: {self.score}", (10, 550), font)
-        # Display correct percentage
-        if self.questions_answered > 0:
-            percentage = int((self.questions_correct / self.questions_answered) * 100)
-            correct_text = f"Correct: {percentage}%"
-        else:
-            correct_text = "Correct: 0%"
-        correct_surface = font.render(correct_text, True, (255, 255, 255))
-        correct_rect = correct_surface.get_rect(center=(400, 560))
-        screen.blit(correct_surface, correct_rect)
-
-    def update(self):
-        if self.state != PLAYING:
-            return  # skip countdowns while paused or on menus
-        if self.state == PLAYING:
-            self.time_left = self.game_length - int(time.time() - self.start_time)
-            if self.time_left <= 0:
-                self.wrong_answers_file = self.save_wrong_answers()
-                self.state = GAME_OVER
-            elif time.time() > self.question_timer:
-                self.questions_answered += 1
-                self.feedback = "Pass"
-                self.feedback_color = (255, 255, 255)
-                self.correct_streak = 0  # Reset streak on pass.
-                # Track passed question as wrong since user didn't answer it
-                wrong_entry = {
-                    'question': self.current_question[0],
-                    'correct_answer': self.current_question[1],
-                    'user_answer': '(No answer - time expired)',
-                    'extra_info': self.current_question[2] if len(self.current_question) > 2 else ''
-                }
-                self.wrong_answers.append(wrong_entry)
-                self.next_question()
-
-    def handle_mouse_click(self, event):
-        if event.type == pygame.MOUSEBUTTONDOWN and event.button == 1:  # Left mouse button
-            if self.state == MENU:
-                start_btn = pygame.Rect(300, 430, 200, 50)
-                clear_btn = pygame.Rect(300, 490, 200, 50)
-                if start_btn.collidepoint(event.pos):
-                    self.game_mode = NORMAL_MODE
-                    self.reset_game()
-                elif clear_btn.collidepoint(event.pos):
-                    self.game_mode = CLEAR_MODE
-                    self.reset_game()
-            elif self.state == GAME_OVER:
-                restart_btn = pygame.Rect(300, 350, 200, 50)
-                menu_btn = pygame.Rect(300, 420, 200, 50)
-                if restart_btn.collidepoint(event.pos):
-                    self.reset_game()
-                elif menu_btn.collidepoint(event.pos):
-                    self.state = MENU
-
-    def handle_input(self, event):
-        if self.state == PLAYING and event.type == pygame.KEYDOWN:
-            if event.key == pygame.K_RETURN or event.key == pygame.K_KP_ENTER:
-                self.questions_answered += 1
-                user_answer = self.user_input.strip().lower()
-                correct_answer = self.current_question[1].strip().lower()
-
-                if user_answer == correct_answer:
-                    # Correct answer.
-                    self.questions_correct += 1
-                    if self.game_mode == CLEAR_MODE:
-                        self.clear_mode_correct += 1
-                        # Remove question from available questions
-                        if self.current_question in self.available_questions:
-                            self.available_questions.remove(self.current_question)
-                    self.correct_streak += 1
-                    multiplier = 1
-                    if self.correct_streak >= 3:
-                        multiplier = self.correct_streak - 1
-                    self.score += 10 * multiplier
-                    self.feedback = f"Correct x{multiplier}"
-                    self.feedback_color = (0, 255, 0)
-                    if self.sound_enabled:
-                        self.sound_manager.play_correct()
-                else:
-                    # Check similarity for partial credit
-                    similarity = calculate_similarity(user_answer, correct_answer)
-                    if similarity >= 0.9:   #Modify this to increase or decrease similarity check
-                        # Close enough - award half points
-                        self.questions_correct += 0.5  # Half credit for statistics
-                        if self.game_mode == CLEAR_MODE:
-                            self.clear_mode_correct += 1
-                            # Remove question from available questions for partial credit too
-                            if self.current_question in self.available_questions:
-                                self.available_questions.remove(self.current_question)
-                        self.correct_streak += 1
-                        multiplier = 1
-                        if self.correct_streak >= 3:
-                            multiplier = self.correct_streak - 1
-                        half_points = int((10 * multiplier) / 2)
-                        self.score += half_points
-                        self.feedback = f"Close! - half points! x{multiplier}"
-                        self.feedback_color = (255, 225, 0)  # Yellow for partial credit
-                        if self.sound_enabled:
-                            self.sound_manager.play_correct()
-                    else:
-                        self.feedback = "Incorrect"
-                        self.feedback_color = (255, 0, 0)
-                        self.correct_streak = 0
-                        # Track wrong answer
-                        wrong_entry = {
-                            'question': self.current_question[0],
-                            'correct_answer': self.current_question[1],
-                            'user_answer': self.user_input.strip(),
-                            'extra_info': self.current_question[2] if len(self.current_question) > 2 else ''
-                        }
-                        self.wrong_answers.append(wrong_entry)
-                        if self.sound_enabled:
-                            self.sound_manager.play_wrong()
-                self.next_question()
-            elif event.key == pygame.K_BACKSPACE:
-                self.user_input = self.user_input[:-1]
-            else:
-                self.user_input += event.unicode
-
-    def draw(self):
-        screen.fill((0, 0, 0))
-        # Display F11 (Sound) and F12 (Learning Mode) toggles in the top-right.
-        sound_text = f"F11 > Sound: {'On' if self.sound_enabled else 'Off'}"
-        lm_text = f"F12 > Learning Mode: {'On' if self.learning_mode else 'Off'}"
-        sound_surface = font.render(sound_text, True, (255, 255, 255))
-        lm_surface = font.render(lm_text, True, (255, 255, 255))
-        sound_rect = sound_surface.get_rect(topright=(790, 10))
-        lm_rect = lm_surface.get_rect(topright=(790, 40))
-        screen.blit(sound_surface, sound_rect)
-        screen.blit(lm_surface, lm_rect)
-        pause_text = "F9 > Pause/Resume"
-        pause_surface = font.render(pause_text, True, (255, 255, 255))
-        pause_rect = pause_surface.get_rect(topright=(790, 70))
-        screen.blit(pause_surface, pause_rect)
-
-        if self.state == MENU:
-            draw_text(screen, "A+ Typing Game", (300, 100), font)
-            draw_text(screen, "Game Length (s):", (300, 270), font)
-            self.input_box_game_length.draw(screen)
-            draw_text(screen, "Question Time (s):", (300, 340), font)
-            self.input_box_question_time.draw(screen)
-            button((300, 430, 200, 50), "Start Game", screen, font)
-            button((300, 490, 200, 50), "Clear Mode", screen, font)
-        elif self.state == PLAYING:
-            draw_wrapped_text(screen, f"{self.current_question[0]}", (40, 200), font)
-            draw_text(screen, f"> {self.user_input}", (40, 300), font)
-            self.draw_all_information()
-        elif self.state == PAUSED:
-            draw_text(screen, f"> {self.user_input}", (40, 300), font)
-            self.draw_all_information()
-            overlay = pygame.Surface((800, 600), pygame.SRCALPHA)
-            overlay.fill((0, 0, 0, 150))
-            screen.blit(overlay, (0, 0))
-            pause_msg = font.render("PAUSED – press F9 to resume", True, (255, 255, 0))
-            rect = pause_msg.get_rect(center=(400, 300))
-            screen.blit(pause_msg, rect)
-        elif self.state == GAME_OVER:
-            draw_text(screen, "Game Over", (335, 200), font)
-            draw_text(screen, f"Final Score: {self.score}", (315, 250), font)
-            if self.game_mode == CLEAR_MODE:
-                clear_stats_text = f"Questions Completed: {self.clear_mode_correct}"
-                draw_text(screen, clear_stats_text, (280, 280), font)
-            if hasattr(self, 'wrong_answers_file') and self.wrong_answers_file:
-                msg = f"Questions missed written to: {os.path.basename(self.wrong_answers_file)}"
-                draw_wrapped_text(screen, msg, (200, 130), font, color=(255, 255, 0), max_width=700)
-            button((300, 350, 200, 50), "Restart", screen, font)
-            button((300, 420, 200, 50), "Main Menu", screen, font)
-
-
 # Main loop
-game = TypingGame(sound_manager)
 running = True
 while running:
     screen.fill((0, 0, 0))
     for event in pygame.event.get():
         if event.type == pygame.QUIT:
             running = False
-
         if event.type == pygame.KEYDOWN and event.key == pygame.K_F9:
-            if game.state == PLAYING:
+            if game.state == "playing":
                 game.pause()
-            elif game.state == PAUSED:
+            elif game.state == "paused":
                 game.resume()
-
-        # Toggle Sound with F11.
         if event.type == pygame.KEYDOWN and event.key == pygame.K_F11:
             game.sound_enabled = not game.sound_enabled
-
-        # Toggle Learning Mode with F12.
         if event.type == pygame.KEYDOWN and event.key == pygame.K_F12:
             game.learning_mode = not game.learning_mode
-
-        # Handle mouse clicks for buttons
         game.handle_mouse_click(event)
-        
-        # Handle input box events in the MENU state.
-        if game.state == MENU:
-            # Handle tabbing between input boxes.
+        if game.state == "menu":
             if event.type == pygame.KEYDOWN and event.key == pygame.K_TAB:
                 if game.input_box_game_length.active:
                     game.input_box_game_length.active = False
@@ -433,23 +70,18 @@ while running:
                     game.input_box_game_length.text = ""
                     game.input_box_game_length.txt_surface = font.render(game.input_box_game_length.text, True,(0, 0, 0))
                     game.input_box_game_length.color = game.input_box_game_length.color_active
-
                 else:
                     game.input_box_game_length.active = True
                     game.input_box_game_length.text = ""
                     game.input_box_game_length.txt_surface = font.render(game.input_box_game_length.text, True,(0, 0, 0))
                     game.input_box_game_length.color = game.input_box_game_length.color_active
-                continue  # Skip further processing of the Tab key.
-
+                continue
             game.input_box_game_length.handle_event(event)
             game.input_box_question_time.handle_event(event)
-        # Handle gameplay input in the PLAYING state.
-        if game.state == PLAYING:
+        if game.state == "playing":
             game.handle_input(event)
-
     game.update()
     game.draw()
     pygame.display.flip()
     clock.tick(60)
-
 pygame.quit()
